renamed "typecheck\: to "semantic"
